Crate buck2_data

Modules

action_error

Nested message and enum types in ActionError.

action_error_diagnostics

Nested message and enum types in ActionErrorDiagnostics.

action_execution_end

Nested message and enum types in ActionExecutionEnd.

action_key

Nested message and enum types in ActionKey.

action_key_owner

analysis_end

Nested message and enum types in AnalysisEnd.

analysis_stage_start

Nested message and enum types in AnalysisStageStart.

analysis_start

Nested message and enum types in AnalysisStart.

buck_event

Nested message and enum types in BuckEvent.

buckconfig_component

Nested message and enum types in BuckconfigComponent.

command_end

Nested message and enum types in CommandEnd.

command_execution

Nested message and enum types in CommandExecution.

command_execution_kind

Nested message and enum types in CommandExecutionKind.

command_invalidation_info

Nested message and enum types in CommandInvalidationInfo.

command_start

Nested message and enum types in CommandStart.

config_file

Nested message and enum types in ConfigFile.

critical_path_entry2

Nested message and enum types in CriticalPathEntry2.

deferred_preparation_stage_start

Nested message and enum types in DeferredPreparationStageStart.

device_metadata

Nested message and enum types in DeviceMetadata.

dynamic_lambda_owner

Nested message and enum types in DynamicLambdaOwner.

dynamic_lambda_start

Nested message and enum types in DynamicLambdaStart.

error

error_report

Nested message and enum types in ErrorReport.

executor_stage_start

Nested message and enum types in ExecutorStageStart.

instant_event

Nested message and enum types in InstantEvent.

local_stage

Nested message and enum types in LocalStage.

materializer_command

Nested message and enum types in MaterializerCommand.

re_platform

Nested message and enum types in RePlatform.

re_stage

Nested message and enum types in ReStage.

record_event

Nested message and enum types in RecordEvent.

serialize_duration_as_micros

serialize_timestamp

span_end_event

Nested message and enum types in SpanEndEvent.

span_start_event

Nested message and enum types in SpanStartEvent.

starlark_user_metadata_value

Nested message and enum types in StarlarkUserMetadataValue.

structured_error

Nested message and enum types in StructuredError.

test_discovery

Nested message and enum types in TestDiscovery.

test_result

Nested message and enum types in TestResult.

Structs

AcquireLocalResource

ActionError

ActionErrorDiagnostics

Either the produced ActionSubErrors, or the error that occurred when invoking the error handler

ActionErrorHandlerExecutionEnd

ActionErrorHandlerExecutionStart

ActionErrorLocation

ActionErrorLocations

Wrapper around ActionErrorLocation so we can differentiate between null and empty locations

ActionExecutionEnd

ActionExecutionStart

The beginning of execution for a particular action.

ActionKey

A unique key identifying a particular action.

ActionName

A name for a particular action, suitable for offline analytics and user display. ActionNames are unique within the execution of a particular target.

ActionOutput

ActionSubError

Error category produced by the error handler function provided by the rule author.

ActionSubErrors

Wrapper around ActionSubError so we can use oneof in ActionErrorDiagnostics

AggregatedBuildMetrics

These use double so it’s convenient to use for both the normal metrics and the amortized case. For the size of values that we see, the precision is fine.

AllTargetsBuildMetrics

Aggregated metrics across all targets in a build.

AnalysisEnd

AnalysisProfile

AnalysisResolveQueriesEnd

AnalysisResolveQueriesStart

AnalysisStageEnd

AnalysisStageStart

AnalysisStart

AnonTarget

AqueryCommandEnd

AqueryCommandStart

AuditCommandEnd

AuditCommandStart

BuckEvent

A single event originating from Buck. Its semantics depend on the data member.

BuckconfigComponent

BuckconfigInputValues

BuildArtifact

BuildCommandEnd

BuildCommandStart

BuildGraphExecutionInfo

Sent once per build.

BuildGraphStats

Record event sent directly to scribe.

BuildResult

BuildTarget

BxlCommandEnd

BxlCommandStart

BxlDiceInvocationEnd

BxlDiceInvocationStart

BxlEnsureArtifactsEnd

BxlEnsureArtifactsStart

BxlExecutionEnd

BxlExecutionStart

BxlFunctionKey

A bxl function key, which is a bxl function and its args

BxlFunctionLabel

A bxl function label, which is its defining file and the function name

CQueryCommandEnd

CQueryCommandStart

CacheHit

CacheQuery

CacheUploadEnd

CacheUploadStart

CellHasNewConfigs

CleanCommandEnd

CleanCommandStart

CleanStaleResult

CleanStaleStats

ClientMetadata

CommandCriticalEnd

Marks the exit of the CommandCriticalStart event, such that the command has left the critical section.

CommandCriticalStart

An event that marks that the command has start execution and synced it’s state. The period of time between CommandStart and CommandCriticalStart is the time spent synchronizing changes and waiting for concurrent commands to finish.

CommandEnd

CommandExecution

Serialization of CommandExecutionReport

CommandExecutionDetails

CommandExecutionError

NOTE: This is an empty message. When this is returned as an error, the relevant execution details are in the reports field.

CommandExecutionKind

CommandExecutionMetadata

Serialization of CommandExecutionMetadata

CommandExecutionStats

CommandInvalidationInfo

CommandOptions

CommandOutputsMissing

CommandReport

CommandStart

An event that marks the beginning of a command.

CommandTimedOut

CompleteCommandEnd

CompleteCommandStart

ComputeDetailedAggregatedMetricsEnd

ComputeDetailedAggregatedMetricsStart

ConcurrentCommands

ConfigFile

ConfigValue

Configuration

A configuration, identified by its full name.

ConfigurationCreated

ConfigurationWithConstraints

A configuration with all the constraints associated with it.

ConfiguredTargetLabel

A configured target label, which is a target label plus a configuration.

ConfiguredTargetsCommandEnd

ConfiguredTargetsCommandStart

ConnectToInstallerEnd

ConnectToInstallerStart

ConsoleMessage

/ A message that should be printed to the user, generally via stderr / / The formatting used is up to the user, and may have things like ANSI / control characters removed. / / Generally, one should rely on properly typed events, rather than raw / console events if additional formatting is required.

ConsolePreferences

ConsoleWarning

/ A message that should be printed to the user as a warning in yellow color, / generally via stderr

Constraint

CpuCounter

CqueryUniverseBuildEnd

CqueryUniverseBuildStart

CreateOutputSymlinksEnd

CreateOutputSymlinksStart

CriticalPathEntry

CriticalPathEntry2

DaemonShutdown

DebugAdapterCommandSnapshot

DebugAdapterSnapshot

The DebugAdapterSnapshot is sent to all commands when a debugger is attached to buck, it contains imformation about the currently paused threads.

DebugAdapterStoppedEval

DeferredEvaluationEnd

DeferredPreparationStageEnd

DeferredPreparationStageStart

DepFileUploadEnd

DepFileUploadStart

DetailedAggregatedMetrics

Detailed metrics aggregated to top-level targets and for the overall build.

DeviceMetadata

DiceBlockConcurrentCommandEnd

DiceBlockConcurrentCommandStart

DiceCleanupEnd

DiceCleanupStart

DiceCriticalSectionEnd

DiceCriticalSectionStart

DiceEqualityCheck

DiceKeyState

DiceStateSnapshot

DiceStateUpdateEnd

DiceStateUpdateStart

DiceSynchronizeSectionEnd

DiceSynchronizeSectionStart

DocsCommandEnd

DocsCommandStart

DynamicLambdaOwner

DynamicLambdaStart

EndOfTestResults

EnvironmentEntry

ErrorReport

ExclusiveCommandWaitEnd

ExclusiveCommandWaitStart

ExecutorStageEnd

ExecutorStageStart

ExpandExternalCellsCommandEnd

ExpandExternalCellsCommandStart

ExplainCommandEnd

ExplainCommandStart

FakeEnd

FakeStart

FileStatusCommandEnd

FileStatusCommandStart

FileWatcherEnd

FileWatcherEvent

FileWatcherStart

FileWatcherStats

Stats for a file watching (e.g. Watchman) update. We ignore events that are outside of the cells we care about, which is why events_total and events_processed might differ.

FreshInstance

GlobalExternalConfig

InstallCommandEnd

InstallCommandStart

InstallEventInfoEnd

InstallEventInfoStart

InstallFinished

InstantEvent

An event that represents a single point in time.

Invocation

Header in binary event log.

InvocationRecord

This is the origin for every sample in buck2_builds scuba table It’s sent from the client to Scribe at the end of each invocation

IoProviderInfo

LoadBuildFileEnd

LoadBuildFileStart

LoadPackageEnd

LoadPackageStart

LocalCommand

/ A representation of a command that we executed locally.

LocalExecute

LocalMaterializeInputs

LocalPrepareOutputDirs

LocalQueued

LocalStage

Location

LspCommandEnd

LspCommandStart

MatchDepFilesEnd

MatchDepFilesStart

MaterializationEnd

MaterializationStart

MaterializeCommandEnd

MaterializeCommandStart

MaterializeFailedInputs

MaterializeRequestedArtifactEnd

MaterializeRequestedArtifactStart

The beginning of materialization for the output of a target requested, inclusive of all dependent artifacts it might recursively request to materialize.

MaterializedArtifacts

MaterializerCommand

Events capturing commands to the deferred materializer. Currently only enabled with -c buck2.verbose_materializer_event_log=true. Events are reported by the command processor, so they are logged when they are processed by the materializer, not when they are sent to the materializer

MaterializerStateInfo

NetworkInterfaceStats

NoActiveDiceState

NoopEvent

OmittedLocalCommand

ParsedTargetPatterns

Unambiguous provider patterns that this command is requesting be brought up-to-date. These patterns are not exactly the same as the patterns directly provided to the buck2 command-line; rather, these patterns have been disambiguated using the client’s working directory and cell configuration.

PersistEventLogSubprocess

PrepareAction

PreviousCommandWithMismatchedConfig

ProcessedErrorReport

Identical to ErrorReport, but with the tags converted to strings. This makes it easy to change those enums without having to worry about uses of this protobuf definition outside of buck2. Currently used in the invocation record and nowhere else.

ProfileCommandEnd

ProfileCommandStart

QueryCommandEnd

QueryCommandStart

QuickUnstableE2eData

This should only be used as a mechanism to easily get information into the event log for e2e tests.

RageResult

Result of invoking buck2 rage

ReAfterAction

ReBeforeAction

ReDownload

ReExecute

RePlatform

ReQueue

ReQueueAcquiringDependencies

ReQueueCancelled

ReQueueNoWorkerAvailable

ReQueueOverQuota

ReStage

ReUnknown

ReUploadEnd

ReUploadMetrics

ReUploadStart

Wraps the RE upload stage of execution to report useful data on network traffic. Eventually all the executor stage span definitions should be moved to dedicated spans.

ReWorkerDownload

ReWorkerUpload

RecordEvent

ReleaseLocalResourcesEnd

ReleaseLocalResourcesStart

RemoteCommand

A representation of a command we executed remotely.

RemoteCommandDetails

RemoteExecutionSessionCreated

RestartConfiguration

SetupLocalResourcesEnd

SetupLocalResourcesStart

SharedTaskEnd

SharedTaskStart

Snapshot

A snapshot of current system state, with useful info.

SoftError

SpanCancelled

Returned when a Span is dropped before terminating.

SpanEndEvent

An event that ends a span.

SpanStartEvent

An event that begins a span.

SpanStats

StarlarkCommandEnd

StarlarkCommandStart

StarlarkDebugAttachCommandEnd

StarlarkDebugAttachCommandStart

StarlarkFailNoStacktrace

StarlarkUserEvent

StarlarkUserMetadataDictValue

StarlarkUserMetadataListValue

StarlarkUserMetadataValue

StdoutStreamingOutput

/ A message that should be streaming printed to the user via stdout

StructuredError

An error that needs to be reported. This may be a panic, or a soft error.

SubscriptionCommandEnd

SubscriptionCommandStart

SystemInfo

TagEvent

An arbitrary tag on the command. Usually used to mark A/B experiments

TargetCfg

Event sent during build commands providing information about global configuration options.

TargetLabel

A target label, consisting of a package and a name.

TargetPattern

A target pattern, representing a set of targets. See <<https://buck2.build/docs/concepts/target_pattern>> for Buck’s documentation on target patterns.

TargetsCommandEnd

TargetsCommandStart

TestCommandEnd

TestCommandStart

TestDiscovery

An event capturing information from the test discovery phase. Test discovery includes sending a summary of the current testing session. For a given target, we also report when we discover its tests.

TestDiscoveryEnd

TestDiscoveryStart

TestResult

TestRunEnd

TestRunStart

TestSessionInfo

At the beginning of discovery, the test orchestrator will advertise some information about the session

TestSuite

We report all of the tests for a particular target simultaneously

TopLevelTargetCriticalPath

TopLevelTargetMetrics

Aggregated metrics associated with a top-level target.

TraceIoCommandEnd

TraceIoCommandStart

TypedMetadata

UnixSystemStats

UntrackedFile

VersionControlRevision

WorkerCommand

/ A representation of a command executed by a local worker.

WorkerExecute

WorkerInit

WorkerInitCommand

WorkerQueued

WorkerWait

Enums

ActionExecutionKind

The kinds of ways an action can be executed by buck2.

ActionKind

CacheHitType

CacheType

CleanStaleResultKind

DaemonWasStartedReason

Why current command started buckd.

FileWatcherEventType

FileWatcherKind

FileWatcherProvider

InvocationOutcome

MaterializationMethod

NetworkKind

SchedulingMode

TestStatus

Traits

ToProtoMessage

Trait for things that can be converted into protobuf messages, for ease of emitting events. There are many core Buck types that are represented in the Daemon API that use this trait to ease conversion.