buck2_data::command_start

Enum Data

pub enum Data {

    Build(BuildCommandStart),
    Targets(TargetsCommandStart),
    Query(QueryCommandStart),
    Cquery(CQueryCommandStart),
    Test(TestCommandStart),
    Audit(AuditCommandStart),
    Docs(DocsCommandStart),
    Clean(CleanCommandStart),
    Aquery(AqueryCommandStart),
    Install(InstallCommandStart),
    Materialize(MaterializeCommandStart),
    Profile(ProfileCommandStart),
    Bxl(BxlCommandStart),
    Lsp(LspCommandStart),
    FileStatus(FileStatusCommandStart),
    Starlark(StarlarkCommandStart),
    Subscribe(SubscriptionCommandStart),
    Trace(TraceIoCommandStart),
    Ctargets(ConfiguredTargetsCommandStart),
    StarlarkDebugAttach(StarlarkDebugAttachCommandStart),
    Explain(ExplainCommandStart),
    ExpandExternalCell(ExpandExternalCellsCommandStart),
    Complete(CompleteCommandStart),

}

Variants

Build(BuildCommandStart)

Targets(TargetsCommandStart)

Query(QueryCommandStart)

Cquery(CQueryCommandStart)

Test(TestCommandStart)

Audit(AuditCommandStart)

Docs(DocsCommandStart)

Clean(CleanCommandStart)

Aquery(AqueryCommandStart)

Install(InstallCommandStart)

Materialize(MaterializeCommandStart)

Profile(ProfileCommandStart)

Bxl(BxlCommandStart)

Lsp(LspCommandStart)

FileStatus(FileStatusCommandStart)

Starlark(StarlarkCommandStart)

Subscribe(SubscriptionCommandStart)

Trace(TraceIoCommandStart)

Ctargets(ConfiguredTargetsCommandStart)

StarlarkDebugAttach(StarlarkDebugAttachCommandStart)

Explain(ExplainCommandStart)

ExpandExternalCell(ExpandExternalCellsCommandStart)

Complete(CompleteCommandStart)

Implementations

impl Data

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Data>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.