buck2_data::error

Enum ErrorTag

#[repr(i32)]
pub enum ErrorTag {

    UnusedDefaultTag = 0,
    Input = 900,
    Tier0 = 901,
    Environment = 902,
    StarlarkFail = 1,
    StarlarkError = 100,
    StarlarkStackOverflow = 102,
    StarlarkInternal = 103,
    StarlarkValue = 104,
    StarlarkFunction = 105,
    StarlarkScope = 106,
    StarlarkParser = 107,
    StarlarkNativeInput = 108,
    WatchmanTimeout = 2,
    WatchmanRequestError = 201,
    WatchmanConnectionError = 202,
    WatchmanConnectionLost = 203,
    WatchmanConnectionDiscovery = 204,
    WatchmanServerError = 205,
    WatchmanResponseError = 206,
    WatchmanMissingField = 207,
    WatchmanDeserialize = 208,
    WatchmanSerialize = 209,
    WatchmanConnect = 210,
    WatchmanRootNotConnectedError = 211,
    WatchmanCheckoutInProgress = 212,
    WatchmanClient = 213,
    NotifyWatcher = 250,
    Http = 3,
    HttpClient = 301,
    HttpForbidden = 3_403,
    HttpServer = 302,
    ClientGrpc = 4,
    DaemonConnect = 5,
    DaemonIsBusy = 501,
    DaemonPreempted = 502,
    DaemonStateInitFailed = 503,
    DaemonStatus = 504,
    DaemonRedirect = 505,
    GrpcResponseMessageTooLarge = 6,
    Visibility = 8,
    ServerStderrEmpty = 11,
    ServerPanicked = 12,
    ServerStackOverflow = 13,
    ServerSegv = 14,
    ServerSigterm = 400,
    ServerSigabrt = 402,
    ServerSigbus = 403,
    ServerMemoryPressure = 401,
    ServerJemallocAssert = 15,
    ServerTransportError = 16,
    ServerStderrUnknown = 19,
    InternalError = 21,
    ProjectMissingPath = 22,
    InterruptedByDaemonShutdown = 23,
    DaemonWontDieFromKill = 24,
    FailedToKill = 8_027,
    NoValidCerts = 25,
    MaterializationError = 26,
    NoBuckRoot = 27,
    Unimplemented = 29,
    MissingTarget = 30,
    ActionMismatchedOutputs = 601,
    ActionMissingOutputs = 602,
    ActionWrongOutputType = 603,
    ActionCommandFailure = 604,
    Install = 200,
    InstallIdMismatch = 2_110,
    InstallerUnknown = 2_101,
    InstallerTier0 = 2_102,
    InstallerEnvironment = 2_103,
    InstallerInput = 2_104,
    Analysis = 7,
    AnyActionExecution = 2_000,
    IoSystem = 1_000,
    IoSource = 1_010,
    IoNotFound = 1_020,
    IoPermissionDenied = 1_021,
    IoBrokenPipe = 1_022,
    IoStorageFull = 1_023,
    IoExecutableFileBusy = 1_024,
    IoConnectionAborted = 1_025,
    IoNotConnected = 1_026,
    IoTimeout = 1_027,
    IoWindowsSharingViolation = 1_028,
    IoReadOnlyFilesystem = 1_029,
    IoInputOutputError = 1_030,
    IoBlockingExecutor = 8_033,
    IoEden = 1_100,
    IoEdenUncategorized = 1_101,
    IoEdenConnectionError = 1_110,
    IoEdenRequestError = 1_111,
    IoEdenMountDoesNotExist = 1_112,
    IoEdenMountNotReady = 1_113,
    IoEdenWin32Error = 1_150,
    IoEdenHresultError = 1_151,
    IoEdenArgumentError = 1_152,
    IoEdenGenericError = 1_153,
    IoEdenMountGenerationChanged = 1_154,
    IoEdenJournalTruncated = 1_155,
    IoEdenCheckoutInProgress = 1_156,
    IoEdenOutOfDateParent = 1_157,
    IoEdenAttributeUnavailable = 1_158,
    IoEdenUnknownField = 1_160,
    IoMaterializerFileBusy = 1_161,
    IoEdenFileNotFound = 1_162,
    IoEdenListMounts = 1_163,
    IoEdenConfigError = 1_164,
    IoEdenVersionError = 1_165,
    IoEdenThriftError = 1_166,
    Sapling = 12_000,
    IoClientBrokenPipe = 1_201,
    ReUnknownTcode = 1_300,
    ReCancelled = 1_301,
    ReUnknown = 1_302,
    ReInvalidArgument = 1_303,
    ReDeadlineExceeded = 1_304,
    ReNotFound = 1_305,
    ReAlreadyExists = 1_306,
    RePermissionDenied = 1_307,
    ReResourceExhausted = 1_308,
    ReFailedPrecondition = 1_309,
    ReAborted = 1_310,
    ReOutOfRange = 1_311,
    ReUnimplemented = 1_312,
    ReInternal = 1_313,
    ReUnavailable = 1_314,
    ReDataLoss = 1_315,
    ReUnauthenticated = 1_316,
    ReExperimentName = 1_317,
    ReClientCrash = 1_320,
    ReCasArtifactWrongNumberOfInputs = 1_321,
    ReCasArtifactWrongNumberOfOutputs = 1_322,
    ReCasArtifactGetDigestExpirationError = 1_323,
    ReCasArtifactInvalidExpiration = 1_324,
    ReCasArtifactExpired = 1_325,
    ReInvalidGetCasResponse = 1_326,
    ReConnection = 1_340,
    ReUserQuota = 1_341,
    DiceDuplicatedChange = 1_400,
    DiceChangedToInvalid = 1_401,
    DiceInjectedKeyGotInvalidation = 1_402,
    DiceCancelled = 1_403,
    DiceUnexpectedCycleGuardType = 1_404,
    DiceDuplicateActivationData = 1_405,
    DiceRejected = 1_406,
    ConfigureAttr = 3_001,
    DepOnlyIncompatible = 3_002,
    CompatibilityError = 3_003,
    TargetIncompatible = 3_004,
    DownloadFileHeadRequest = 4_001,
    DownloadSizeMismatch = 4_002,
    DigestTtlMismatch = 4_103,
    DigestTtlInvalidResponse = 4_104,
    TestDeadlineExpired = 5_001,
    TestOrchestrator = 5_002,
    TestStatus = 5_003,
    TestStatusInvalid = 5_004,
    LocalResourceSetup = 5_005,
    Tpx = 5_500,
    TestExecutor = 5_501,
    CleanStale = 280,
    CleanInterrupt = 28,
    InvalidErrorReport = 6_002,
    InvalidEvent = 6_003,
    InvalidAuthToken = 6_004,
    InvalidDigest = 6_005,
    InvalidDuration = 6_006,
    InvalidUsername = 6_007,
    InvalidAbsPath = 6_008,
    InvalidBuckOutPath = 6_009,
    BuildDeadlineExpired = 7_001,
    BuildSketchError = 7_002,
    SuperConsole = 8_000,
    SuperConsoleInvalidWhitespace = 8_001,
    WorkerInit = 8_500,
    WorkerDirectoryExists = 8_501,
    WorkerCancelled = 8_502,
    EventLogUpload = 8_035,
    EventLogNotOpen = 8_037,
    EventLogEof = 8_039,
    EventLogIndexOutOfBounds = 8_040,
    CacheUploadFailed = 906,
    SymlinkParentMissing = 909,
    DispatcherUnavailable = 912,
    CsvParse = 917,
    CasBlobCountMismatch = 918,
    CleanOutputs = 8_017,
    CopyOutputs = 8_029,
    LogFilter = 8_030,
    WindowsUnsupported = 8_007,
    MallocStats = 8_008,
    Mallctl = 8_038,
    CpuStats = 8_026,
    Bxl = 11_000,
    Certs = 22_000,
    LogCmd = 13_000,
    OfflineArchive = 14_000,
    Profile = 15_000,
    Lsp = 16_000,
    HealthCheck = 17_000,
    Explain = 18_000,
    StarlarkServer = 19_000,
    Interpreter = 20_000,
    KillAll = 21_000,
    Clap = 9_001,
    ClapMatch = 90_011,
    Hex = 9_002,
    Hyper = 9_003,
    Nix = 9_004,
    Prost = 9_005,
    Regex = 9_006,
    RelativePath = 9_007,
    Rusqlite = 9_008,
    Tokio = 9_018,
    Tonic = 9_019,
    Uuid = 9_020,
    SerdeJson = 9_021,
    StdSlice = 9_009,
    StdTime = 9_010,
    StdInfallible = 9_011,
    StdStripPrefix = 9_012,
    ParseNum = 9_013,
    ParseBool = 9_017,
    IntConversion = 9_014,
    StringUtf8 = 9_015,
    StringConversion = 9_016,
    CstringNul = 9_023,
    CrashRequested = 8_022,
    Bail = 8_023,
    UnexpectedNone = 6_001,
    TestOnly = 8_031,

}

Error types are - by design - restricted to being set exactly once at the error definition site. While they are useful, that means that they are enough on their own to represent all the error metadata we want. Until we figure out what exactly a more complete model looks like, error tags a mechanism that offers maximum flexibility - they can be added to any error anywhere.

Feel free to continue using error types if you like.

Variants

UnusedDefaultTag = 0

Same as above

Input = 900

Tier0 = 901

Environment = 902

StarlarkFail = 1

StarlarkError = 100

StarlarkStackOverflow = 102

StarlarkInternal = 103

StarlarkValue = 104

StarlarkFunction = 105

StarlarkScope = 106

StarlarkParser = 107

StarlarkNativeInput = 108

WatchmanTimeout = 2

WatchmanRequestError = 201

WatchmanConnectionError = 202

Taken from watchman_client::Error

WatchmanConnectionLost = 203

WatchmanConnectionDiscovery = 204

WatchmanServerError = 205

WatchmanResponseError = 206

WatchmanMissingField = 207

WatchmanDeserialize = 208

WatchmanSerialize = 209

WatchmanConnect = 210

WatchmanRootNotConnectedError = 211

WatchmanCheckoutInProgress = 212

WatchmanClient = 213

NotifyWatcher = 250

Http = 3

HttpClient = 301

Client error (4xx).

HttpForbidden = 3_403

HttpServer = 302

Server error (5xx).

ClientGrpc = 4

gRPC protocol error between client and server from the client side.

Protocol error (e.g. malformed frame, or too large frame)

Transport error (e.g. connection closed)

Not application error (e.g. bzl file not found)

DaemonConnect = 5

Connect to buckd failed.

DaemonIsBusy = 501

Daemon is running another command.

DaemonPreempted = 502

Daemon was preempted during preemptible command by another command.

DaemonStateInitFailed = 503

Daemon start up failed while initializing DaemonStateData.

DaemonStatus = 504

DaemonRedirect = 505

GrpcResponseMessageTooLarge = 6

Too large gRPC message.

Visibility = 8

visibility, within_view.

ServerStderrEmpty = 11

Server stderr is empty.

ServerPanicked = 12

Server stderr indicates that the server panicked.

ServerStackOverflow = 13

Server stack overflow.

ServerSegv = 14

SEGV.

ServerSigterm = 400

Server received SIGTERM

ServerSigabrt = 402

Server received SIGABRT

ServerSigbus = 403

Server received SIGBUS

ServerMemoryPressure = 401

Server disconnect with no error but memory pressure was detected.

ServerJemallocAssert = 15

Jemalloc assertion failure.

ServerTransportError = 16

Server transport errors. Eg. Channel not open

ServerStderrUnknown = 19

The reason for server failure is unknown.

InternalError = 21

Internal error in buck2. This is a bug.

ProjectMissingPath = 22

Artifact projection to a path that does not exist

InterruptedByDaemonShutdown = 23

The daemon reported that it was shutting down during the execution of this command

DaemonWontDieFromKill = 24

The daemon couldn’t be killed

FailedToKill = 8_027

A process couldn’t be killed

NoValidCerts = 25

No valid internal or VPNless certs could be found

MaterializationError = 26

Build failed during materialization

NoBuckRoot = 27

Could not find buck project root

Unimplemented = 29

Functionality not implemented in Buck2

MissingTarget = 30

Target specified is not found in the package/build file

ActionMismatchedOutputs = 601

ActionMissingOutputs = 602

ActionWrongOutputType = 603

ActionCommandFailure = 604

Install = 200

Errors during buck2 install.

InstallIdMismatch = 2_110

InstallerUnknown = 2_101

InstallerTier0 = 2_102

InstallerEnvironment = 2_103

InstallerInput = 2_104

Analysis = 7

// High level descriptions of the “phase” of the build during which the error occurred

AnyActionExecution = 2_000

IoSystem = 1_000

/// IO SECTION

Indicates that the IO operation went through the standard system interfaces, and not through EdenIO - note that the operation may still have been accessing an Eden mount though

IoSource = 1_010

IO done on a source file in the repo

IoNotFound = 1_020

The nature of the failure, designed after (but not identical to) Rust’s io::ErrorKind

IoPermissionDenied = 1_021

IoBrokenPipe = 1_022

IoStorageFull = 1_023

IoExecutableFileBusy = 1_024

IoConnectionAborted = 1_025

IoNotConnected = 1_026

IoTimeout = 1_027

IoWindowsSharingViolation = 1_028

IoReadOnlyFilesystem = 1_029

IoInputOutputError = 1_030

IoBlockingExecutor = 8_033

IoEden = 1_100

Eden IO Section Indicates that the IO operation went through Eden

IoEdenUncategorized = 1_101

IO Error encountered in Eden that we do not correspond to a specific error

IoEdenConnectionError = 1_110

Failures indicating that Eden Failed to Connect or Mount

IoEdenRequestError = 1_111

IoEdenMountDoesNotExist = 1_112

IoEdenMountNotReady = 1_113

IoEdenWin32Error = 1_150

The underlying cause of request failures, copied from edenfs::EdenError

IoEdenHresultError = 1_151

IoEdenArgumentError = 1_152

IoEdenGenericError = 1_153

IoEdenMountGenerationChanged = 1_154

IoEdenJournalTruncated = 1_155

IoEdenCheckoutInProgress = 1_156

IoEdenOutOfDateParent = 1_157

IoEdenAttributeUnavailable = 1_158

IoEdenUnknownField = 1_160

IoMaterializerFileBusy = 1_161

IoEdenFileNotFound = 1_162

IoEdenListMounts = 1_163

IoEdenConfigError = 1_164

IoEdenVersionError = 1_165

IoEdenThriftError = 1_166

Sapling = 12_000

IoClientBrokenPipe = 1_201

Client IO Broken pipe specifically from client stdio streams

ReUnknownTcode = 1_300

/// Remote Execution

ReCancelled = 1_301

RE TCode values: https://fburl.com/code/1ael5pmz

ReUnknown = 1_302

ReInvalidArgument = 1_303

ReDeadlineExceeded = 1_304

ReNotFound = 1_305

ReAlreadyExists = 1_306

RePermissionDenied = 1_307

ReResourceExhausted = 1_308

ReFailedPrecondition = 1_309

ReAborted = 1_310

ReOutOfRange = 1_311

ReUnimplemented = 1_312

ReInternal = 1_313

ReUnavailable = 1_314

ReDataLoss = 1_315

ReUnauthenticated = 1_316

ReExperimentName = 1_317

ReClientCrash = 1_320

ReCasArtifactWrongNumberOfInputs = 1_321

ReCasArtifactWrongNumberOfOutputs = 1_322

ReCasArtifactGetDigestExpirationError = 1_323

ReCasArtifactInvalidExpiration = 1_324

ReCasArtifactExpired = 1_325

ReInvalidGetCasResponse = 1_326

ReConnection = 1_340

TCodeReasonGroup values: https://fburl.com/code/pvsy2jgv

ReUserQuota = 1_341

DiceDuplicatedChange = 1_400

Dice Errors Value: https://fburl.com/code/5dxzaw41

DiceChangedToInvalid = 1_401

DiceInjectedKeyGotInvalidation = 1_402

DiceCancelled = 1_403

DiceUnexpectedCycleGuardType = 1_404

DiceDuplicateActivationData = 1_405

DiceRejected = 1_406

ConfigureAttr = 3_001

Error during attribute configuration during target configuration.

DepOnlyIncompatible = 3_002

Dep only incompatible error

CompatibilityError = 3_003

Generic compatibility error tag

TargetIncompatible = 3_004

target configuration is incompatible error

DownloadFileHeadRequest = 4_001

Action execution

DownloadSizeMismatch = 4_002

DigestTtlMismatch = 4_103

DigestTtlInvalidResponse = 4_104

TestDeadlineExpired = 5_001

Tests

TestOrchestrator = 5_002

TestStatus = 5_003

TestStatusInvalid = 5_004

LocalResourceSetup = 5_005

Tpx = 5_500

Error from TestExecutor

TestExecutor = 5_501

CleanStale = 280

Errors while cleaning stale artifacts

CleanInterrupt = 28

Clean stale command was interrupted

InvalidErrorReport = 6_002

InvalidEvent = 6_003

InvalidAuthToken = 6_004

InvalidDigest = 6_005

InvalidDuration = 6_006

InvalidUsername = 6_007

InvalidAbsPath = 6_008

InvalidBuckOutPath = 6_009

BuildDeadlineExpired = 7_001

Build errors.

BuildSketchError = 7_002

SuperConsole = 8_000

SuperConsoleInvalidWhitespace = 8_001

WorkerInit = 8_500

Misc external error conversions

WorkerDirectoryExists = 8_501

WorkerCancelled = 8_502

EventLogUpload = 8_035

EventLogNotOpen = 8_037

EventLogEof = 8_039

EventLogIndexOutOfBounds = 8_040

CacheUploadFailed = 906

SymlinkParentMissing = 909

DispatcherUnavailable = 912

CsvParse = 917

CasBlobCountMismatch = 918

CleanOutputs = 8_017

CopyOutputs = 8_029

LogFilter = 8_030

WindowsUnsupported = 8_007

MallocStats = 8_008

Mallctl = 8_038

CpuStats = 8_026

Bxl = 11_000

Misc phases/subsystem tags

Certs = 22_000

LogCmd = 13_000

OfflineArchive = 14_000

Profile = 15_000

Lsp = 16_000

HealthCheck = 17_000

Explain = 18_000

StarlarkServer = 19_000

Interpreter = 20_000

KillAll = 21_000

Clap = 9_001

Misc error conversions

ClapMatch = 90_011

Hex = 9_002

Hyper = 9_003

Nix = 9_004

Prost = 9_005

Regex = 9_006

RelativePath = 9_007

Rusqlite = 9_008

Tokio = 9_018

Tonic = 9_019

Uuid = 9_020

SerdeJson = 9_021

StdSlice = 9_009

StdTime = 9_010

StdInfallible = 9_011

StdStripPrefix = 9_012

ParseNum = 9_013

ParseBool = 9_017

IntConversion = 9_014

StringUtf8 = 9_015

StringConversion = 9_016

CstringNul = 9_023

CrashRequested = 8_022

From buck2 debug crash

Bail = 8_023

From ExitResult::bail

UnexpectedNone = 6_001

None turned into an error. This carries no useful information whatsoever.

TestOnly = 8_031

Only used by tests, should not occur outside test execution.

Implementations

impl ErrorTag

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ErrorTag.

pub fn from_i32(value: i32) -> Option<ErrorTag>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ErrorTag, or None if value is not a valid variant.

impl ErrorTag

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.