buck2_data::instant_event

Enum Data

pub enum Data {

    StructuredError(StructuredError),
    ConsoleMessage(ConsoleMessage),
    BuildGraphInfo(BuildGraphExecutionInfo),
    ReSession(RemoteExecutionSessionCreated),
    TestDiscovery(TestDiscovery),
    TestResult(TestResult),
    Snapshot(Box<Snapshot>),
    DiceStateSnapshot(DiceStateSnapshot),
    TagEvent(TagEvent),
    TargetPatterns(ParsedTargetPatterns),
    DiceEqualityCheck(DiceEqualityCheck),
    NoActiveDiceState(NoActiveDiceState),
    MaterializerStateInfo(MaterializerStateInfo),
    DaemonShutdown(DaemonShutdown),
    RageResult(RageResult),
    ConsolePreferences(ConsolePreferences),
    IoProviderInfo(IoProviderInfo),
    StarlarkFailNoStacktrace(StarlarkFailNoStacktrace),
    DebugAdapterSnapshot(DebugAdapterSnapshot),
    RestartConfiguration(RestartConfiguration),
    UntrackedFile(UntrackedFile),
    StarlarkUserEvent(StarlarkUserEvent),
    ComandOptions(CommandOptions),
    ConcurrentCommands(ConcurrentCommands),
    PersistEventLogSubprocess(PersistEventLogSubprocess),
    ActionError(ActionError),
    ConsoleWarning(ConsoleWarning),
    MaterializerCommand(MaterializerCommand),
    CleanStaleResult(CleanStaleResult),
    InstallFinished(InstallFinished),
    SystemInfo(SystemInfo),
    VersionControlRevision(VersionControlRevision),
    TargetCfg(TargetCfg),
    UnstableE2eData(QuickUnstableE2eData),
    EndOfTestResults(EndOfTestResults),
    ConfigurationCreated(ConfigurationCreated),
    BuckconfigInputValues(BuckconfigInputValues),
    CellHasNewConfigs(CellHasNewConfigs),
    PreviousCommandWithMismatchedConfig(PreviousCommandWithMismatchedConfig),
    DetailedAggregatedMetrics(DetailedAggregatedMetrics),
    StreamingOutput(StdoutStreamingOutput),

}

Variants

StructuredError(StructuredError)

ConsoleMessage(ConsoleMessage)

BuildGraphInfo(BuildGraphExecutionInfo)

ReSession(RemoteExecutionSessionCreated)

TestDiscovery(TestDiscovery)

Test discovery occurs as an instant event because we know when it starts, but we do not know when it is complete. Instead, we are informed of individual tests that are discovered.

TestResult(TestResult)

Test result is an instant event for similar reasons as TestDiscovery. It corresponds with the completion of an individual test.

Snapshot(Box<Snapshot>)

DiceStateSnapshot(DiceStateSnapshot)

TagEvent(TagEvent)

A tag to mark on the scuba tables

TargetPatterns(ParsedTargetPatterns)

Sent when the target pattern gets resolved to update the invocation info

DiceEqualityCheck(DiceEqualityCheck)

Check DICE equality in critical section

NoActiveDiceState(NoActiveDiceState)

When we can execute DICE ops immediately since there was no active state

MaterializerStateInfo(MaterializerStateInfo)

Info about materializer state

DaemonShutdown(DaemonShutdown)

Notify the client that the daemon is shutting down.

RageResult(RageResult)

Information collected by buck2 rage command

ConsolePreferences(ConsolePreferences)

Notify the client that the daemon would like to update the rendering.

IoProviderInfo(IoProviderInfo)

StarlarkFailNoStacktrace(StarlarkFailNoStacktrace)

Stacktrace from the fail_no_stacktrace() call.

DebugAdapterSnapshot(DebugAdapterSnapshot)

Snapshot of current debug adapter state. Only sent when a debugger is attached.

RestartConfiguration(RestartConfiguration)

This is notionally reading from .buckconfig at startup, but since that’s done by the daemon, it needs to come through this way.

UntrackedFile(UntrackedFile)

Unexpected file found in buck-out/<isolation_dir>/gen during a clean –stale run, not found in materializer state

StarlarkUserEvent(StarlarkUserEvent)

For user-defined instant events. Currently only used in BXL.

ComandOptions(CommandOptions)

Log options that are command-level and processed by the daemon.

ConcurrentCommands(ConcurrentCommands)

Concurrent active commands.

PersistEventLogSubprocess(PersistEventLogSubprocess)

Info coming from the buck2 debug persist-event-log subprocess

ActionError(ActionError)

An action error encountered during the build

ConsoleWarning(ConsoleWarning)

MaterializerCommand(MaterializerCommand)

CleanStaleResult(CleanStaleResult)

InstallFinished(InstallFinished)

SystemInfo(SystemInfo)

VersionControlRevision(VersionControlRevision)

TargetCfg(TargetCfg)

UnstableE2eData(QuickUnstableE2eData)

Just something for us to be able to easily propagate out internal information. Used for testing.

EndOfTestResults(EndOfTestResults)

/ Notification that execution of all tests for a given command is / finished.

ConfigurationCreated(ConfigurationCreated)

Event sent when creating a new configuration.

BuckconfigInputValues(BuckconfigInputValues)

Tracks values of external buckconfigs

CellHasNewConfigs(CellHasNewConfigs)

PreviousCommandWithMismatchedConfig(PreviousCommandWithMismatchedConfig)

Previous command, but only logged if external configs changed

DetailedAggregatedMetrics(DetailedAggregatedMetrics)

Detailed aggregated metrics. Expensive to compute, so not always available.

StreamingOutput(StdoutStreamingOutput)

Event that used for emitting streaming std output

Implementations

impl Data

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Data>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.