buck2_data::span_end_event

Enum Data

pub enum Data {

    Command(CommandEnd),
    ActionExecution(Box<ActionExecutionEnd>),
    Analysis(AnalysisEnd),
    AnalysisResolveQueries(AnalysisResolveQueriesEnd),
    Load(LoadBuildFileEnd),
    ExecutorStage(ExecutorStageEnd),
    TestDiscovery(TestDiscoveryEnd),
    TestEnd(TestRunEnd),
    SpanCancelled(SpanCancelled),
    FileWatcher(FileWatcherEnd),
    FinalMaterialization(MaterializeRequestedArtifactEnd),
    AnalysisStage(AnalysisStageEnd),
    MatchDepFiles(MatchDepFilesEnd),
    LoadPackage(LoadPackageEnd),
    SharedTask(SharedTaskEnd),
    CacheUpload(Box<CacheUploadEnd>),
    CreateOutputSymlinks(CreateOutputSymlinksEnd),
    CommandCritical(CommandCriticalEnd),
    InstallEventInfo(InstallEventInfoEnd),
    DiceStateUpdate(DiceStateUpdateEnd),
    Materialization(MaterializationEnd),
    DiceCriticalSection(DiceCriticalSectionEnd),
    DiceBlockConcurrentCommand(DiceBlockConcurrentCommandEnd),
    DiceSynchronizeSection(DiceSynchronizeSectionEnd),
    DiceCleanup(DiceCleanupEnd),
    ExclusiveCommandWait(ExclusiveCommandWaitEnd),
    DeferredPreparationStage(DeferredPreparationStageEnd),
    DeferredEvaluation(DeferredEvaluationEnd),
    BxlExecution(BxlExecutionEnd),
    BxlDiceInvocation(BxlDiceInvocationEnd),
    ReUpload(ReUploadEnd),
    ConnectToInstaller(ConnectToInstallerEnd),
    LocalResources(SetupLocalResourcesEnd),
    ReleaseLocalResources(ReleaseLocalResourcesEnd),
    BxlEnsureArtifacts(BxlEnsureArtifactsEnd),
    ActionErrorHandlerExecution(ActionErrorHandlerExecutionEnd),
    CqueryUniverseBuild(CqueryUniverseBuildEnd),
    DepFileUpload(DepFileUploadEnd),
    ComputeDetailedAggregatedMetrics(ComputeDetailedAggregatedMetricsEnd),
    Fake(FakeEnd),

}

Variants

Command(CommandEnd)

ActionExecution(Box<ActionExecutionEnd>)

Analysis(AnalysisEnd)

AnalysisResolveQueries(AnalysisResolveQueriesEnd)

Load(LoadBuildFileEnd)

ExecutorStage(ExecutorStageEnd)

TestDiscovery(TestDiscoveryEnd)

TestEnd(TestRunEnd)

SpanCancelled(SpanCancelled)

FileWatcher(FileWatcherEnd)

FinalMaterialization(MaterializeRequestedArtifactEnd)

AnalysisStage(AnalysisStageEnd)

MatchDepFiles(MatchDepFilesEnd)

LoadPackage(LoadPackageEnd)

SharedTask(SharedTaskEnd)

CacheUpload(Box<CacheUploadEnd>)

CreateOutputSymlinks(CreateOutputSymlinksEnd)

CommandCritical(CommandCriticalEnd)

InstallEventInfo(InstallEventInfoEnd)

DiceStateUpdate(DiceStateUpdateEnd)

Materialization(MaterializationEnd)

DiceCriticalSection(DiceCriticalSectionEnd)

DiceBlockConcurrentCommand(DiceBlockConcurrentCommandEnd)

DiceSynchronizeSection(DiceSynchronizeSectionEnd)

DiceCleanup(DiceCleanupEnd)

ExclusiveCommandWait(ExclusiveCommandWaitEnd)

DeferredPreparationStage(DeferredPreparationStageEnd)

DeferredEvaluation(DeferredEvaluationEnd)

BxlExecution(BxlExecutionEnd)

BxlDiceInvocation(BxlDiceInvocationEnd)

ReUpload(ReUploadEnd)

ConnectToInstaller(ConnectToInstallerEnd)

LocalResources(SetupLocalResourcesEnd)

ReleaseLocalResources(ReleaseLocalResourcesEnd)

BxlEnsureArtifacts(BxlEnsureArtifactsEnd)

ActionErrorHandlerExecution(ActionErrorHandlerExecutionEnd)

CqueryUniverseBuild(CqueryUniverseBuildEnd)

DepFileUpload(DepFileUploadEnd)

ComputeDetailedAggregatedMetrics(ComputeDetailedAggregatedMetricsEnd)

Fake(FakeEnd)

Used in Buck unit tests.

Implementations

impl Data

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Data>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.