buck2_data

Struct ProcessedErrorReport

pub struct ProcessedErrorReport {
    pub tier: Option<i32>,
    pub message: String,
    pub telemetry_message: Option<String>,
    pub source_location: Option<String>,
    pub tags: Vec<String>,
    pub best_tag: Option<String>,
    pub sub_error_categories: Vec<String>,
    pub category_key: Option<String>,
    pub category: Option<String>,
    pub source_area: Option<String>,
}

Identical to ErrorReport, but with the tags converted to strings. This makes it easy to change those enums without having to worry about uses of this protobuf definition outside of buck2. Currently used in the invocation record and nowhere else.

Fields
tier: Option<i32>message: Stringtelemetry_message: Option<String>source_location: Option<String>tags: Vec<String>best_tag: Option<String>

buck2_error crate has logic of selecting the most interesting error tag among all error tags. This is such tag.

sub_error_categories: Vec<String>category_key: Option<String>category: Option<String>source_area: Option<String>
Implementations

impl ProcessedErrorReport

pub fn tier(&self) -> ErrorTier

Returns the enum value of tier, or the default if the field is unset or set to an invalid enum value.

pub fn set_tier(&mut self, value: ErrorTier)

Sets tier to the provided enum value.

pub fn telemetry_message(&self) -> &str

Returns the value of telemetry_message, or the default value if telemetry_message is unset.

pub fn source_location(&self) -> &str

Returns the value of source_location, or the default value if source_location is unset.

pub fn best_tag(&self) -> &str

Returns the value of best_tag, or the default value if best_tag is unset.

pub fn category_key(&self) -> &str

Returns the value of category_key, or the default value if category_key is unset.

pub fn category(&self) -> &str

Returns the value of category, or the default value if category is unset.

pub fn source_area(&self) -> &str

Returns the value of source_area, or the default value if source_area is unset.