buck2_data

Struct Snapshot

pub struct Snapshot {

    pub buck2_rss: Option<u64>,
    pub buck2_max_rss: u64,
    pub buck2_user_cpu_us: u64,
    pub buck2_system_cpu_us: u64,
    pub blocking_executor_io_queue_size: u64,
    pub re_download_bytes: u64,
    pub re_upload_bytes: u64,
    pub re_uploads_started: u32,
    pub re_uploads_finished_successfully: u32,
    pub re_uploads_finished_with_error: u32,
    pub re_downloads_started: u32,
    pub re_downloads_finished_successfully: u32,
    pub re_downloads_finished_with_error: u32,
    pub re_action_cache_started: u32,
    pub re_action_cache_finished_successfully: u32,
    pub re_action_cache_finished_with_error: u32,
    pub re_executes_started: u32,
    pub re_executes_finished_successfully: u32,
    pub re_executes_finished_with_error: u32,
    pub re_materializes_started: u32,
    pub re_materializes_finished_successfully: u32,
    pub re_materializes_finished_with_error: u32,
    pub re_write_action_results_started: u32,
    pub re_write_action_results_finished_successfully: u32,
    pub re_write_action_results_finished_with_error: u32,
    pub re_get_digest_expirations_started: u32,
    pub re_get_digest_expirations_finished_successfully: u32,
    pub re_get_digest_expirations_finished_with_error: u32,
    pub io_in_flight_copy: u32,
    pub io_in_flight_symlink: u32,
    pub io_in_flight_hardlink: u32,
    pub io_in_flight_mk_dir: u32,
    pub io_in_flight_read_dir: u32,
    pub io_in_flight_read_dir_eden: u32,
    pub io_in_flight_rm_dir: u32,
    pub io_in_flight_rm_dir_all: u32,
    pub io_in_flight_stat: u32,
    pub io_in_flight_stat_eden: u32,
    pub io_in_flight_chmod: u32,
    pub io_in_flight_read_link: u32,
    pub io_in_flight_remove: u32,
    pub io_in_flight_rename: u32,
    pub io_in_flight_read: u32,
    pub io_in_flight_write: u32,
    pub io_in_flight_canonicalize: u32,
    pub io_in_flight_eden_settle: u32,
    pub daemon_uptime_s: u64,
    pub malloc_bytes_active: Option<u64>,
    pub malloc_bytes_allocated: Option<u64>,
    pub used_disk_space_bytes: Option<u64>,
    pub host_cpu_usage_system_ms: Option<u64>,
    pub host_cpu_usage_user_ms: Option<u64>,
    pub dice_key_count: u64,
    pub dice_currently_active_key_count: u64,
    pub dice_active_transaction_count: u32,
    pub deferred_materializer_queue_size: u64,
    pub sink_successes: Option<u64>,
    pub sink_failures: Option<u64>,
    pub sink_failures_invalid_request: Option<u64>,
    pub sink_failures_unauthorized: Option<u64>,
    pub sink_failures_rate_limited: Option<u64>,
    pub sink_failures_pushed_back: Option<u64>,
    pub sink_failures_enqueue_failed: Option<u64>,
    pub sink_failures_internal_error: Option<u64>,
    pub sink_failures_timed_out: Option<u64>,
    pub sink_failures_unknown: Option<u64>,
    pub sink_buffer_depth: Option<u64>,
    pub sink_dropped: Option<u64>,
    pub sink_bytes_written: Option<u64>,
    pub network_interface_stats: HashMap<String, NetworkInterfaceStats>,
    pub http_download_bytes: u64,
    pub deferred_materializer_declares: u64,
    pub deferred_materializer_declares_reused: u64,
    pub unix_system_stats: Option<UnixSystemStats>,
    pub zdb_download_queries: u64,
    pub zdb_download_bytes: u64,
    pub zdb_upload_queries: u64,
    pub zdb_upload_bytes: u64,
    pub zgateway_download_queries: u64,
    pub zgateway_download_bytes: u64,
    pub zgateway_upload_queries: u64,
    pub zgateway_upload_bytes: u64,
    pub manifold_download_queries: u64,
    pub manifold_download_bytes: u64,
    pub manifold_upload_queries: u64,
    pub manifold_upload_bytes: u64,
    pub hedwig_download_queries: u64,
    pub hedwig_download_bytes: u64,
    pub hedwig_upload_queries: u64,
    pub hedwig_upload_bytes: u64,
    pub local_cache_hits_files: i64,
    pub local_cache_hits_bytes: i64,
    pub local_cache_misses_files: i64,
    pub local_cache_misses_bytes: i64,
    pub local_cache_hits_files_from_memory_cache: i64,
    pub local_cache_hits_files_from_filesystem_cache: i64,
    pub local_cache_lookups: i64,
    pub local_cache_lookup_latency_microseconds: i64,
    pub this_event_client_delay_ms: Option<i64>,
    pub client_cpu_percents: Option<u32>,

}

A snapshot of current system state, with useful info.

Fields
buck2_rss: Option<u64>

Resident set size in bytes of the buck2 daemon. Does not include subprocesses (e.g. local actions).

buck2_max_rss: u64

Maximum resident set size in bytes of the buck2 daemon. Does not include subprocesses (e.g. local actions).

buck2_user_cpu_us: u64

User CPU time of buck2 daemon, not including subprocesses.

buck2_system_cpu_us: u64

System CPU time of buck2 daemon, not including subprocesses.

blocking_executor_io_queue_size: u64

Queue size of the blocking executor.

re_download_bytes: u64re_upload_bytes: u64re_uploads_started: u32re_uploads_finished_successfully: u32re_uploads_finished_with_error: u32re_downloads_started: u32re_downloads_finished_successfully: u32re_downloads_finished_with_error: u32re_action_cache_started: u32re_action_cache_finished_successfully: u32re_action_cache_finished_with_error: u32re_executes_started: u32re_executes_finished_successfully: u32re_executes_finished_with_error: u32re_materializes_started: u32re_materializes_finished_successfully: u32re_materializes_finished_with_error: u32re_write_action_results_started: u32re_write_action_results_finished_successfully: u32re_write_action_results_finished_with_error: u32re_get_digest_expirations_started: u32re_get_digest_expirations_finished_successfully: u32re_get_digest_expirations_finished_with_error: u32io_in_flight_copy: u32

I/O operations in progress.

io_in_flight_symlink: u32io_in_flight_hardlink: u32io_in_flight_mk_dir: u32io_in_flight_read_dir: u32io_in_flight_read_dir_eden: u32io_in_flight_rm_dir: u32io_in_flight_rm_dir_all: u32io_in_flight_stat: u32io_in_flight_stat_eden: u32io_in_flight_chmod: u32io_in_flight_read_link: u32io_in_flight_remove: u32io_in_flight_rename: u32io_in_flight_read: u32io_in_flight_write: u32io_in_flight_canonicalize: u32io_in_flight_eden_settle: u32daemon_uptime_s: u64

Time passed since buck2 daemon was started

malloc_bytes_active: Option<u64>

https://jemalloc.net/jemalloc.3.html (stats.active) Total number of bytes in active pages allocated by the application, is greater or equal to stats.allocated.

malloc_bytes_allocated: Option<u64>

(stats.allocated) Total number of bytes allocated by the application

used_disk_space_bytes: Option<u64>

Total number of bytes of used disk space on machine

host_cpu_usage_system_ms: Option<u64>

Host system CPU time since command start.

host_cpu_usage_user_ms: Option<u64>

Host user CPU time since command start.

dice_key_count: u64dice_currently_active_key_count: u64

the number of keys actively present in the per transaction cache

dice_active_transaction_count: u32deferred_materializer_queue_size: u64sink_successes: Option<u64>

Sink write statistics; counts of sink statistics taken at this snapshot. Cumulative count of messages that were successfully emitted.

sink_failures: Option<u64>

Cumulative count of messages that failed to be emitted.

sink_failures_invalid_request: Option<u64>

These are only stored in snapshot events, but not written to scuba.

sink_failures_unauthorized: Option<u64>sink_failures_rate_limited: Option<u64>sink_failures_pushed_back: Option<u64>sink_failures_enqueue_failed: Option<u64>sink_failures_internal_error: Option<u64>sink_failures_timed_out: Option<u64>sink_failures_unknown: Option<u64>sink_buffer_depth: Option<u64>

Current number of messages queued up to be submitted.

sink_dropped: Option<u64>

Cumulative count of messages that were dropped (i.e. not even processed).

sink_bytes_written: Option<u64>network_interface_stats: HashMap<String, NetworkInterfaceStats>

Network statistics for “interesting” network interfaces.

http_download_bytes: u64deferred_materializer_declares: u64deferred_materializer_declares_reused: u64unix_system_stats: Option<UnixSystemStats>zdb_download_queries: u64zdb_download_bytes: u64zdb_upload_queries: u64zdb_upload_bytes: u64zgateway_download_queries: u64zgateway_download_bytes: u64zgateway_upload_queries: u64zgateway_upload_bytes: u64manifold_download_queries: u64manifold_download_bytes: u64manifold_upload_queries: u64manifold_upload_bytes: u64hedwig_download_queries: u64hedwig_download_bytes: u64hedwig_upload_queries: u64hedwig_upload_bytes: u64local_cache_hits_files: i64

Local cache hits and misses stats

local_cache_hits_bytes: i64local_cache_misses_files: i64local_cache_misses_bytes: i64local_cache_hits_files_from_memory_cache: i64

Local cache hits per cache layer

local_cache_hits_files_from_filesystem_cache: i64local_cache_lookups: i64local_cache_lookup_latency_microseconds: i64

Client side metrics.

this_event_client_delay_ms: Option<i64>

Delay between time snapshot is created and time it is received by the client.

client_cpu_percents: Option<u32>

Client average user + system CPU time since last snapshot processed.

Implementations

impl Snapshot

pub fn malloc_bytes_active(&self) -> u64

Returns the value of malloc_bytes_active, or the default value if malloc_bytes_active is unset.

pub fn malloc_bytes_allocated(&self) -> u64

Returns the value of malloc_bytes_allocated, or the default value if malloc_bytes_allocated is unset.

pub fn used_disk_space_bytes(&self) -> u64

Returns the value of used_disk_space_bytes, or the default value if used_disk_space_bytes is unset.

pub fn buck2_rss(&self) -> u64

Returns the value of buck2_rss, or the default value if buck2_rss is unset.

pub fn host_cpu_usage_system_ms(&self) -> u64

Returns the value of host_cpu_usage_system_ms, or the default value if host_cpu_usage_system_ms is unset.

pub fn host_cpu_usage_user_ms(&self) -> u64

Returns the value of host_cpu_usage_user_ms, or the default value if host_cpu_usage_user_ms is unset.

pub fn sink_successes(&self) -> u64

Returns the value of sink_successes, or the default value if sink_successes is unset.

pub fn sink_failures(&self) -> u64

Returns the value of sink_failures, or the default value if sink_failures is unset.

pub fn sink_buffer_depth(&self) -> u64

Returns the value of sink_buffer_depth, or the default value if sink_buffer_depth is unset.

pub fn sink_dropped(&self) -> u64

Returns the value of sink_dropped, or the default value if sink_dropped is unset.

pub fn sink_bytes_written(&self) -> u64

Returns the value of sink_bytes_written, or the default value if sink_bytes_written is unset.

pub fn sink_failures_invalid_request(&self) -> u64

Returns the value of sink_failures_invalid_request, or the default value if sink_failures_invalid_request is unset.

pub fn sink_failures_unauthorized(&self) -> u64

Returns the value of sink_failures_unauthorized, or the default value if sink_failures_unauthorized is unset.

pub fn sink_failures_rate_limited(&self) -> u64

Returns the value of sink_failures_rate_limited, or the default value if sink_failures_rate_limited is unset.

pub fn sink_failures_pushed_back(&self) -> u64

Returns the value of sink_failures_pushed_back, or the default value if sink_failures_pushed_back is unset.

pub fn sink_failures_enqueue_failed(&self) -> u64

Returns the value of sink_failures_enqueue_failed, or the default value if sink_failures_enqueue_failed is unset.

pub fn sink_failures_internal_error(&self) -> u64

Returns the value of sink_failures_internal_error, or the default value if sink_failures_internal_error is unset.

pub fn sink_failures_timed_out(&self) -> u64

Returns the value of sink_failures_timed_out, or the default value if sink_failures_timed_out is unset.

pub fn sink_failures_unknown(&self) -> u64

Returns the value of sink_failures_unknown, or the default value if sink_failures_unknown is unset.

pub fn this_event_client_delay_ms(&self) -> i64

Returns the value of this_event_client_delay_ms, or the default value if this_event_client_delay_ms is unset.

pub fn client_cpu_percents(&self) -> u32

Returns the value of client_cpu_percents, or the default value if client_cpu_percents is unset.